This commit is adopted from r1039: Fixed titler bug: occasional full intensity flash at last frame
Bug: Titles configured to fade out sometimes flashed the title at full
intensity at the last frame of the effect range.

The problem was a boundary condition of the alpha calculation such that if
the last frame of the effect was exactly aligned with a frame boundary,
alpha would be set to 1 instead of 0.

Also changed some integer truncation to rounding, although this wasn't
related to the bug.

[j6t: use lroundf(...) instead of (int)roundf(...)]

Patch by Craig Lawson.

Original commit by Johannes Sixt <johannes.sixt@telecom.at>

Adopted line-numbers and left out one line to make the fix fit for cin4