This commit is adopted from r1017: Fix curved masks boundaries if start and end point are very close.
When the start and end points of a mask bound were only a few pixels apart,
then a Bezier curve between them would be interpolated by too few points.
To fix this, we use the same algorithm as in the mask engine, which is a
forward differencing method.

Original commit by Johannes Sixt <johannes.sixt@telecom.at>

Adopted line-numbers to make fix fit for cin4