WIP: renovate SVG plugin: add helper application rawc-convert
The SVG via Inkscape plugin of cinelerra seems to be broken
by all newer Inkscape versions. The reason therefor is that
--cinelerra-export-file="foobar" is not supported anymore.

However, this plugin relies on a file in the homebrewn RAWC
format. This helper application converts a PNG graphic into
such a RAWC image by using "convert" from the imagemagick
suite and adding a RAWC header at the beginning of the file.