WIP: add "colorgrade" plugin
This plugin is far from finished and is subject to change.
Please do not use it for important long-run projects as later versions are likely to be
incompatible.

The current implementation is basically a modification of cinelerra's histogram plugin.
Currently only RGB(A)888(8) is supported for input/output.

This plugin can set the hue of a pixel dependend on its original hue and its value.

It is planned to extend and generalize this to any combination of hue, saturation and value
as input/output parameter.

RGB to HSV conversion as described in http://zach.in.tu-clausthal.de/teaching/cg1_0708/folien/13_color_3_4up.pdf
HSV to RGB conversion as described at http://www.cs.rit.edu/~ncs/color/t_convert.html