Eliminate useless indexes from YUV4MPEG2 file loader (Patch by Monty)
The YUV4MPEG2 loader was building a from-scratch transient index that
required reading every byte in the raw file on every call to open.
Open is called often, not just when the project is loaded.  If the
loader already had an index, it threw it away.  Indexes existed only
in memory, nothing was saved. This means that, for example, every undo
required every y4m file in the project to rebuild indexes from
scratch.  On the movie I was working on, that was 20 minutes as it
re-read just short of a terabyte of video.

There is no practical reason for the index. In yuv4mpeg2 every frame
(including the frame header) is a fixed size.  Although the spec
references theoretical support for a variable number of variably sized
flags per frame header, the only actual flags defined are for
interlacing order, which must appear in every frame and are all the
same size.  Thus, yuv4mpeg2 frame size does not vary.

Remove the YUV4MPEG indexing code completely; rework open to determine
file header and frame size in a robust fashion; rework seek to jump
directly to a computed offset.

This change, in my current project, reduces 'open' and 'undo' time
from ten minutes to instantaneous.  It also eliminates the bug where
compositor playback hits a new track and the entire UI becomes
unresponsive until the open call is finished rebuilding the transient
index.